refactoring and import rules fixes
