+++ src/Pages/Diets/DietCard/DietCard.module.scss
