--- src/Pages/Routine/Routine.module.scss
@@ -1,51 +1,0 @@
-@use '../../config/color/configColors.scss' as Colors;
-@use '../../config/fontSize/configFontSize.scss' as FontSize;
-@use '../../config/mixins' as Mixins;
-
-.page {
-	width: 90%;
-	margin: 0 5%;
-	font-family: Montserrat;
-	h1,
-	h2,
-	h3,
-	h4,
-	h5,
-	p,
-	span {
-		margin: 0;
-		height: fit-content;
-	}
-
-	.header {
-		color: white;
-		margin: 30px 0;
-		display: flex;
-		width: fit-content;
-
-		hr {
-			margin: 0 10px;
-		}
-	}
-
-	.body {
-        display:flex ;
-        flex-direction:column;
-        align-items: center;
-
-		.exerciseContainer {
-            align-self: center;
-            width: 80%;
-            display: flex;
-            flex-wrap: wrap;
-            gap: 20px;
-            align-items: center;
-		}
-
-        #start{
-            @include Mixins.blue-button;
-            width: 100px;
-            height: 40px;
-        }
-	}
-}
