--- src/Pages/RoutineForm/RoutineForm.module.scss
@@ -1,3 +1,0 @@
-.form{
-    color: white;
-}
