+++ src/Pages/Routines/CircuitPill/CircuitPill.jsx
@@ -1,6 +1,6 @@
 import style from './CircuitPill.module.scss';
-import run from '../../assets/images/run.svg';
-import clock from '../../assets/images/clock.svg';
+import run from '../../../assets/images/run.svg';
+import clock from '../../../assets/images/clock.svg';
 
 export default function CircuitPill({ id, name, length, duration }) {
 	//id queda guardado para routing a una ruta de detalles
