+++ src/Pages/Routines/RoutineBox/RoutineBox.jsx
@@ -1,7 +1,7 @@
 import style from './RoutineBox.module.scss';
-import AvatarGroup from '../AvatarGroup/AvatarGroup.jsx';
-import CircuitPill from '../../Pages/Routines/CircuitPill/CircuitPill.jsx';
-import DietPill from '../DietPill/DietPill.jsx';
+import AvatarGroup from '../../../components/AvatarGroup/AvatarGroup.jsx';
+import CircuitPill from '../CircuitPill/CircuitPill.jsx';
+import DietPill from '../DietPill/DietPill';
 
 export default function RoutineBox({ name, users, circuits, diet }) {
     return (
